// front/src/main/web/src/app/api/model/nonRecurringScheduledTaskTriggerDto.ts
/**
 * i18n Tool
 * Web API of the i18n tool
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ScheduledTaskTriggerDto } from './scheduledTaskTriggerDto';

/**
 * Non-recurring trigger asking the single execution of a task on a certain date.
 */
export interface NonRecurringScheduledTaskTriggerDto extends ScheduledTaskTriggerDto {
  /**
   * Date when the task must be executed.
   */
  startTime: Date;
}
export namespace NonRecurringScheduledTaskTriggerDto {}