// front/src/main/web/src/app/api/model/recurringScheduledTaskTriggerDto.ts
/**
 * i18n Tool
 * Web API of the i18n tool
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ScheduledTaskTriggerDto } from './scheduledTaskTriggerDto';

/**
 * Recurring trigger based on a CRON expression.
 */
export interface RecurringScheduledTaskTriggerDto extends ScheduledTaskTriggerDto { 
    /**
     * CRON expression specifying the periodicity of the task.
     */
    cronExpression: string;
}
export namespace RecurringScheduledTaskTriggerDto {
}