// front/src/main/web/src/app/api/model/scheduledTaskDefinitionDto.ts
/**
 * i18n Tool
 * Web API of the i18n tool
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { LocalizedStringDto } from './localizedStringDto';
import { NonRecurringScheduledTaskTriggerDto } from './nonRecurringScheduledTaskTriggerDto';
import { RecurringScheduledTaskTriggerDto } from './recurringScheduledTaskTriggerDto';

/**
 * Definition of a scheduled task.
 */
export interface ScheduledTaskDefinitionDto { 
    /**
     * Unique id of this kind of persistent task.
     */
    id: string;
    name: LocalizedStringDto;
    description: LocalizedStringDto;
    trigger: NonRecurringScheduledTaskTriggerDto | RecurringScheduledTaskTriggerDto;
    /**
     * Flag indicating whether the specified definition is enabled and can be executed.
     */
    enabled: boolean;
    /**
     * Last time the associated task was executed.
     */
    lastExecutionTime?: Date;
    /**
     * The date when the next execution will occur.
     */
    nextExecutionTime?: Date;
}